Parce options with getopt
With getopt we don't need to seperate the options anymore. So you can
run 'pachist -qn' instead of 'pachist -q -n'.

We also get better error handeling with options.

One side effect is that the number variable will have to have a space
between itself and the option infront of it. That is, you will have to
run 'pachist -q 5' instead of 'pachist -q5'.